Split out product types, fix initialisers, tests and UISandbox.
